refactor(lib): 提取 plugins 域 7 模块 + migration（Task 5a）
- lib/plugin-registry-data.ts: COMPANION_PLUGINS 38 项清单 + 更新源表（pluginUpdateSources 改参数注入解耦 settings）
- lib/plugin-copy.ts: 内容戳记跳过式拷贝家族（COPY_STAMP/pluginCopyEntries/pluginStampOf/copyPluginPackage）
- lib/plugin-manager-core.ts: 启停/移除/恢复/汇总 + js-yaml dsh 方言 + imagePasteSave
- lib/plugins.ts: syncCompanionPlugins（市场同名迁移预检/bundle 重复行清理/孤儿行清理/坏行 heal）+ skills/preset 同步 + healProfileModules + managedPackageNames/restoreKeptArtifacts
- lib/market-ops.ts: 排队任务（快照→执行→allowBuilds 重试→回填→超时强杀）
- lib/migration.ts: 共享 profile 一次性迁移 + 皮肤选择落位 + 临时目录告警
- lib/session-heal.ts: V4 运行时补丁三落点
- 垫片：preset-sync/builtin-collision/patch-row-heal/plugin-manager-state/profile-module-heal/plugin-updater + scripts 3 个 + js-yaml ambient（集中 declare module 会拉源文件，拆独立 d.ts）
- patch-row-heal/better-sidebar-bundle 断言指向 lib 模块（语义等价）
- main.js 3105→1923 行
- 验证：npm test 450/450 全绿、typecheck 零错、check-syntax 通过
