+++ dsh-desktop/main.js
@@ -87,6 +87,24 @@ const {
   showMainWindow, createTray, trayHintOnce, closeToTrayEnabled, setCloseToTray,
   getExitAction, setExitAction, askExitAction, showAbout, repoUrls,
 } = require('./lib/tray.js');
+// Task 5 提取的模块：插件同步/复制/清单/管理/市场任务/迁移/运行时补丁。
+const { COMPANION_PLUGINS, pluginUpdateSources, builtinPluginSourceDir } = require('./lib/plugin-registry-data.js');
+const { readJsonFile, copyPluginPackage } = require('./lib/plugin-copy.js');
+const { SKINS_DIR, healProfileModules, managedPackageNames, restoreKeptArtifacts, syncBundledSkills, syncCompanionPlugins } = require('./lib/plugins.js');
+const {
+  loadDshYamlDialect: _loadDshYamlDialect, pluginManagerReadPatch: _pluginManagerReadPatch,
+  pluginManagerPackageDescription, pluginManagerCollect, pluginManagerResolveName,
+  removedPluginIds, saveRemovedPluginIds: _saveRemovedPluginIds, restoreCompanionPlugin: _restoreCompanionPlugin,
+  pluginManagerSetRemoved, pluginManagerSetEnabled, imagePasteSave,
+} = require('./lib/plugin-manager-core.js');
+const { processPendingMarketOps } = require('./lib/market-ops.js');
+const { warnTempRun, migrateFromSharedWebProfile, extractPatchRowIds, removePatchRowsById, applyLegacySkinChoice } = require('./lib/migration.js');
+const { applySessionManageFix } = require('./lib/session-heal.js');
+// bridge 更新：plugins 域函数已迁入 lib，装配指向新模块导出。
+bridge.processPendingMarketOps = processPendingMarketOps;
+bridge.syncCompanionPlugins = syncCompanionPlugins;
+bridge.healProfileModules = healProfileModules;
+bridge.restoreKeptArtifacts = restoreKeptArtifacts;
 // bridge 更新：窗口/托盘函数已迁入 lib，装配指向新模块导出。
 bridge.showMainWindow = showMainWindow;
 bridge.showBox = showBox;
@@ -96,10 +114,6 @@ bridge.trayHintOnce = trayHintOnce;
 // 跨域注入点装配（lib/bridge.ts）：函数声明提升，此处顶层引用安全。
 bridge.ensureGuard = ensureGuard;
 bridge.handleBootFailure = handleBootFailure;
-bridge.processPendingMarketOps = processPendingMarketOps;
-bridge.syncCompanionPlugins = syncCompanionPlugins;
-bridge.healProfileModules = healProfileModules;
-bridge.restoreKeptArtifacts = restoreKeptArtifacts;
 
 const APP_VERSION = app.getVersion();
 const AUTO_UPDATE_INTERVAL_MS = 6 * 60 * 60 * 1000; // every 6 hours
@@ -638,7 +652,7 @@ function notifyPluginUpdates(updatable) {
 async function runPluginUpdateCheck(manual) {
   if (state.quitting) return;
   const ctx = updCtx();
-  const sources = pluginUpdateSources();
+  const sources = pluginUpdateSources(removedPluginIds());
   if (sources.length === 0) return;
   if (!manual && !pluginUpdater.dueForCheck(ctx, Date.now())) return;
   let list;
@@ -970,7 +984,7 @@ function registerChromeIpc() {
     if (!state.mainWindow || event.sender !== state.mainWindow.webContents) return null;
     try {
       const ctx = updCtx();
-      const list = await pluginUpdater.checkPluginUpdates(ctx, pluginUpdateSources(), {
+      const list = await pluginUpdater.checkPluginUpdates(ctx, pluginUpdateSources(removedPluginIds()), {
         force: !!force,
         profileDirP: desktopProfileDir(),
       });
@@ -987,7 +1001,7 @@ function registerChromeIpc() {
 
   ipcMain.handle('dsh:plugin-update', async (event, { id } = {}) => {
     if (!state.mainWindow || event.sender !== state.mainWindow.webContents) return { ok: false, error: 'unauthorized' };
-    const source = pluginUpdateSources().find((s) => s.id === String(id));
+    const source = pluginUpdateSources(removedPluginIds()).find((s) => s.id === String(id));
     if (!source) return { ok: false, error: '未知或不可更新的内置插件: ' + String(id) };
     try {
       const res = await pluginUpdater.applyBuiltinPluginUpdate(updCtx(), source, {
@@ -1322,1062 +1336,8 @@ function startBalanceLoop() {
   if (state.balanceTimer.unref) state.balanceTimer.unref();
 }
 
-// ---------------------------------------------------------------------------
-// 配套 dsh 插件同步（注入 web profile：余额小部件 + 文件更改追踪/还原 + 皮肤）
-// ---------------------------------------------------------------------------
-
-const COMPANION_PLUGINS = [
-  { id: 'balance', name: '@deepseek-ai/dsh-balance' },
-  { id: 'file-changes', name: '@deepseek-ai/dsh-file-changes' },
-  { id: 'client-file-changes', name: '@deepseek-ai/dsh-client-file-changes' },
-  { id: 'terminal', name: '@deepseek-ai/dsh-terminal' },
-  // 社区插件市场（awesome-dsh-plugin.com 目录）：内置分发，替换早期 npm 检索版市场。
-  { id: 'dsh-market-plugin', name: '@sanqi-normal/dsh-webui-market-plugin', dir: 'dsh-webui-market' },
-  { id: 'skin-switch', name: '@deepseek-ai/dsh-skin-switch' },
-  { id: 'easy-setup', name: '@deepseek-ai/dsh-easy-setup' },
-  // 社区功能插件（视觉 / 人设 / 长期记忆 / 移动端布局修复）：npm registry
-  // 拉取后随应用内置分发。绝不能写进 profile package.json 依赖 ——
-  // pnpm 安装会 hoist @deepseek-ai 核心包形成模块双实例（Symbol 冲突，
-  // 插件命名空间注册失效，即 "设置命名空间不可用" 故障的根因）。
-  { id: 'tool-vision', name: 'dsh-tool-vision', dir: 'dsh-tool-vision' },
-  // config.path 必须随行写入：v2.0.0 只写了 id+name，而当时插件 schema 的
-  // path 是 required 无默认值，全新安装校验失败拖垮整个插件树（dsh web
-  // 退出码 1，应用持续闪退“启动失败”）。schema 现已带默认值，这里显式
-  // 写 config 是双保险，healSoulMdPatchRow 另负责修复存量坏行。
-  { id: 'soul-md', name: 'dsh-soul-md', dir: 'dsh-soul-md', config: { path: 'soul.md' } },
-  { id: 'tdai-memory', name: 'dsh-tdai-memory', dir: 'dsh-tdai-memory' },
-  { id: 'mobile-fix', name: 'dsh-web-mobile-fix', dir: 'dsh-web-mobile-fix' },
-  // VSCode 风格右侧边栏（文件树 / 编辑器 / 终端 / Git，按会话隔离）。
-  // lib/ 预编译自包含（codemirror、xterm 已内嵌），服务端仅额外依赖
-  // schemastery（已加入 app 闭包，见 package.json）。
-  { id: 'better-sidebar', name: 'dsh-better-sidebar', dir: 'dsh-better-sidebar' },
-  // Trae 风格对话回退：用户消息 hover 出「编辑并回退」，按上一完整回合
-  // 分叉新会话（sessions.fork）并以编辑后内容重发（inputActions）。
-  // 纯客户端实现，host 半边为 no-op。
-  { id: 'message-rewind', name: 'dsh-message-rewind', dir: 'dsh-message-rewind' },
-  // 页面桌宠（npm: dsh-pet 0.1.3）：28 个透明动画的悬浮宠物，即装即用。
-  // assets/ 15MB 播放资源随包分发；peer 依赖全部由 dsh 宿主提供。
-  // V4 关键修复：行必须带 config —— dsh-pet 的 apply 读 config.fullRoot，
-  // 无 config 块的行会让 loader 传 undefined 直接拖垮插件树（v3.1.0 全新
-  // 安装即「启动失败」的根因之一；老用户因市场装过的行带 config 才幸免）。
-  // 值沿用包内 cordis.patch.yml 的出厂默认。
-  // 默认禁用 —— 需要页面桌宠时在「设置 → 插件 → 管理」或「桌宠」分区开启。
-  { id: 'dsh-pet', name: 'dsh-pet', dir: 'dsh-pet', config: { size: 260, position: 'bottom-right' }, disabled: true },
-  // 第二插件市场 Zat-DSH Engine（GitHub releases 分发，v0.5.0 vendor 自
-  // 源码 tag）：GitHub dsh-plugin topic 检索 + 中文简介 + 国内镜像兜底。
-  // 运行时依赖 zod ^4 由 profiles 闭包（junction 指向 app node_modules，
-  // 内含 zod 4.4.3）解析，无需 vendor。
-  { id: 'zat-market', name: 'zat-dsh-engine', dir: 'zat-dsh-engine' },
-  // 设置页「Skills 与 MCP」分区：Skills 目录浏览（来源徽标/打开目录）+
-  // MCP 服务增删改（读写 profile patch 中的 dsh-mcp-client 行）+ 从
-  // Claude Code / Codex 一键导入 MCP 配置。
-  { id: 'dock-settings', name: 'dsh-dock-settings', dir: 'dsh-dock-settings' },
-  // 外观自定义：字体家族/字号/文字与代码颜色的设置页分区，实时预览，
-  // localStorage 持久化（纯客户端，无宿主半边）。
-  { id: 'font-custom', name: 'dsh-font-custom', dir: 'dsh-font-custom' },
-  // 自动压缩：监听 contextPressure 投影，接近上下文上限（默认 80%）时
-  // 自动向当前会话发送 /compact（dsh 原生命令，压缩事务由内核执行）。
-  { id: 'auto-compact', name: 'dsh-auto-compact', dir: 'dsh-auto-compact' },
-  // 插件保护中心 UI：快照列表/一键回滚/健康检查/事故报告，经桌面壳
-  // IPC（guard:action）驱动 plugin-guard.js 引擎。
-  { id: 'plugin-shield', name: 'dsh-plugin-shield', dir: 'dsh-plugin-shield' },
-  // AI 变更审核（V4，用户建议⑤）：监控官方 fileChanges 投影，手动/自动向
-  // 当前对话发送审核请求，让模型复查自己刚做的改动（正确性/安全性/一致
-  // 性），结论配合「文件」页一键还原。纯客户端实现，host 半边 no-op。
-  { id: 'change-review', name: 'dsh-change-review', dir: 'dsh-change-review' },
-  // —— V4 自上游 dsh_desktop（myYangyunfan）移植的配套插件 ——
-  // 会话浮窗（多窗口分屏）：会话头部「弹出到独立窗口」按钮；窗口由壳层
-  // IPC chrome:float-window / preload 的 __DSH_FLOAT__ 承载。
-  { id: 'float-window', name: '@deepseek-ai/dsh-float-window' },
-  // 对话节点导航条（vlln/dsh-navbar，MIT）：对话区右缘节点串快速跳转
-  // user 消息（悬停预览/点击跳转/滚轮切换）。
-  { id: 'dsh-navbar', name: '@vlln/dsh-navbar', dir: 'dsh-navbar' },
-  // 对话删除与归档管理：会话行菜单「删除对话」+ 设置内归档管理面板。
-  // 前置依赖 scripts/patch-session-manage.js 的官方包运行时补丁
-  // （applySessionManageFix，随启动幂等应用、覆盖 agent overlay）。
-  { id: 'dsh-session-manager', name: 'dsh-session-manager' },
-  // 对话界面微调：隐藏大量工具调用/结果/思考输出（保留每轮最终总结）。
-  { id: 'conversation-tweaks', name: '@deepseek-ai/dsh-conversation-tweaks' },
-  // 自定义注入提示词：整体替换/追加官方 persona，应用到 standard 预设。
-  { id: 'prompt-custom', name: '@deepseek-ai/dsh-prompt-custom' },
-  // 第三方 OpenAI 兼容模型的 reasoning_effort 控件（字段名可自定义）。
-  { id: 'third-party-thinking', name: '@deepseek-ai/dsh-third-party-thinking' },
-  // 侧边临时会话：浮窗追问、不写主会话、多种回答引擎（Ctrl+Shift+S）。
-  { id: 'side-session', name: '@dsh-external/dsh-side-session', dir: 'dsh-side-session' },
-  // 插件启停管理：设置页「插件 → 管理」标签，不重启切换插件启停
-  // （IPC dsh:plugin-list / dsh:plugin-set-enabled，见下方接线）。
-  { id: 'plugin-manager', name: '@deepseek-ai/dsh-plugin-manager' },
-  // 插件选择向导入口（设置页「插件 → 选择向导」分区）：重新打开首次启动的
-  // 内置插件选择向导，按需启用/停用内置插件。纯客户端 UI + 壳层 IPC
-  // （onboard:*），host 半边 no-op；核心插件组内锁定，永不被向导停用。
-  { id: 'plugin-wizard', name: 'dsh-plugin-wizard', dir: 'dsh-plugin-wizard' },
-  // 微信 ClawBot / OpenClaw 桥（openclaw-dsh-bridge v0.7.0，MIT）：设置页
-  // 「ClawBot」栏（扫码绑定微信官方 ClawBot 小程序）+ OpenAI 兼容端点
-  // （/openclaw-bridge/v1/chat/completions）。设置命名空间经 dsh-host-apiproxy
-  // 的 settings.describe 全量暴露（rc.7+ 已移除 WEB_SETTINGS_NAMESPACES 白名单）。
-  { id: 'openclaw-bridge', name: '@deepseek-ai/dsh-openclaw-bridge', dir: 'dsh-openclaw-bridge' },
-  // 崩溃急救/撤销回退（dsh-undo-savepoint，lire1131，MIT）：配置文件 + 插件
-  // 代码树快照、undo/redo、一键安全模式、密钥脱敏 vault。与插件保护中心
-  // （配置面快照）和「文件」还原（会话内改动）互补，覆盖「配置改坏、dsh
-  // 起不来」的急救场景。GitHub 分发锁定拷贝（npm 未发布）。
-  { id: 'dsh-undo', name: 'dsh-undo-savepoint', dir: 'dsh-undo-savepoint' },
-  // 大肥鱼桌宠（dsh-dafeiyu，QCYTSN；代码 MIT、角色素材按 ASSET_LICENSE.md
-  // 随包分发保留署名）：真实会话状态驱动的原生置顶桌宠（空闲/思考/工作/
-  // 等待/完成/错误 六态 + 项目状态卡）。默认开启 —— 可在「设置 → 插件 →
-  // 管理」或「桌宠」分区关闭（含 49MB PyInstaller helper，按需运行）。
-  { id: 'dsh-dafeiyu', name: 'dsh-dafeiyu', dir: 'dsh-dafeiyu' },
-  // 桌宠设置分区（V4.2，dsh-pet-settings）：设置页「桌宠」分区，集中管理
-  // 页面桌宠（dsh-pet 开关，重启生效）与大肥鱼桌面伴侣（启用/角色大小/
-  // 空闲微动作频率/减少动态，走 dsh-dafeiyu config 端点即时生效）。
-  { id: 'dsh-pet-settings', name: 'dsh-pet-settings', dir: 'dsh-pet-settings' },
-  // 峰谷价格卫士（dsh-offpeak，christophersmith2737-commits，MIT）：DeepSeek
-  // 峰谷定价（2026-08-17 起）高峰时段（北京时间 9-12 / 14-18 点）在发送前
-  // 拦截提醒，可一键继续或定时到闲时价自动执行（浏览器不在线也会到点
-  // 执行）。与余额小部件互补（事前拦截 vs 事后显示）；程序化提交
-  // （auto-compact / 变更审核 / 消息回退 / openclaw 桥）不被拦截。
-  // 可在「设置 → 插件 → 管理」关闭。
-  { id: 'offpeak', name: 'dsh-offpeak', dir: 'dsh-offpeak' },
-  // 拖入文件到对话（V4.1，用户建议）：文本/代码文件拖进输入框自动注入
-  // 内容（上限 256KB），图片/二进制/超大文件注入路径提示配合
-  // inspect_image 与文件工具；纯客户端实现（host 半边 no-op）。
-  { id: 'file-drop', name: 'dsh-file-drop', dir: 'dsh-file-drop' },
-  // 设置页左侧边栏自定义（V4.1，用户建议）：设置面板导航底部「自定义
-  // 边栏」按钮，按需显示/隐藏与排序 settings.section 导航项，
-  // localStorage 持久化，默认全显；纯客户端实现（host 半边 no-op）。
-  { id: 'settings-nav-custom', name: 'dsh-settings-nav-custom', dir: 'dsh-settings-nav-custom' },
-  // 设置页「常规」页内高级选项折叠（V4.2，用户建议）：按行标题关键词把
-  // 低频选项行（外观/语言/权限预设等）收进底部「高级选项」折叠组，
-  // localStorage 持久化展开状态；纯客户端实现（host 半边 no-op）。
-  { id: 'settings-groups', name: 'dsh-settings-groups', dir: 'dsh-settings-groups' },
-  // 图片粘贴发送（V4.2，用户建议）：Ctrl/Cmd+V 粘贴剪贴板图片 → 保存到
-  // 临时目录 → 注入完整路径提示（配合 inspect_image 视觉工具）；纯客户端
-  // 实现（host 半边 no-op，仅用受控 IPC dsh:image-paste-save）。
-  { id: 'image-paste', name: 'dsh-image-paste', dir: 'dsh-image-paste' },
-];
-
-// ---------------------------------------------------------------------------
-// 内置插件上游更新源（V4.3，plugin-updater.js 消费）：
-//
-// 只登记「上游仍在 npm / GitHub 发布」的社区插件 —— 内置分发的副本可以
-// 跟随上游修复而更新。EAC 独占插件（package.json 标记 private，如
-// dsh-balance / dsh-terminal）绝不登记；zat-market 自带 selfupdate 不登记。
-// 运行时 npm 404（未上架/改名）优雅降级为「无上游」，绝不阻塞。
-// ---------------------------------------------------------------------------
-const PLUGIN_UPDATE_SOURCES = {
-  'tool-vision': { npm: 'dsh-tool-vision' },
-  'soul-md': { npm: 'dsh-soul-md' },
-  'tdai-memory': { npm: 'dsh-tdai-memory' },
-  'dsh-pet': { npm: 'dsh-pet' },
-  'better-sidebar': { npm: 'dsh-better-sidebar' },
-  'dsh-navbar': { npm: '@vlln/dsh-navbar' },
-  'mobile-fix': { npm: 'dsh-web-mobile-fix' },
-  'offpeak': { npm: 'dsh-offpeak' },
-  'dsh-market-plugin': { npm: '@sanqi-normal/dsh-webui-market-plugin' },
-  'dsh-session-manager': { npm: 'dsh-session-manager' },
-  // GitHub 分发（npm 未发布）：dsh-undo-savepoint。
-  'dsh-undo': { github: 'lire1131/dsh-undo-savepoint' },
-};
-
-/** 把内置插件表 + 更新源注册表合并成 plugin-updater 的 sources 输入。 */
-function pluginUpdateSources() {
-  const removed = removedPluginIds();
-  const out = [];
-  for (const p of COMPANION_PLUGINS) {
-    const update = PLUGIN_UPDATE_SOURCES[p.id];
-    if (!update) continue;
-    if (removed.has(p.id)) continue;
-    const dirName = p.dir || (p.name.includes('/') ? p.name.split('/').pop() : p.name);
-    const assetsDir = path.join(__dirname, 'assets', 'plugins', dirName);
-    if (!fs.existsSync(path.join(assetsDir, 'package.json'))) continue;
-    out.push({ id: p.id, name: p.name, assetsDir, update });
-  }
-  return out;
-}
-
-/** 内置插件当前生效的源目录：覆盖层（已更新版本）优先，资产版本回退。 */
-function builtinPluginSourceDir(dirName) {
-  const assets = path.join(__dirname, 'assets', 'plugins', dirName);
-  const overlay = path.join(state.userDataDir, 'builtin-plugin-updates', dirName);
-  if (!fs.existsSync(path.join(overlay, 'package.json'))) return assets;
-  if (!fs.existsSync(path.join(assets, 'package.json'))) return overlay;
-  // 覆盖层版本 >= 资产版本才优先：应用自身升级后，新资产自动接管覆盖层。
-  const vOverlay = pluginUpdater.versionOfDir(overlay);
-  const vAssets = pluginUpdater.versionOfDir(assets);
-  if (vOverlay && vAssets && updater.compareVersions(vOverlay, vAssets) < 0) return assets;
-  return overlay;
-}
-
-// 皮肤包目录：assets/skins/<id>/。每个皮肤是一个完整的 dsh client 插件包
-// （package.json + lib/ + skin.json + LICENSE/NOTICE），随桌面端分发；
-// 默认全部以 disabled: true 注册（不启用任何皮肤），由「设置 → 皮肤」切换。
-const SKINS_DIR = path.join(__dirname, 'assets', 'skins');
-
-function readJsonFile(file) {
-  try { return JSON.parse(fs.readFileSync(file, 'utf8')); } catch { return null; }
-}
-
-// 拷贝一个插件包目录到 profile node_modules（按包名 scope 落位，幂等）。
-// 除运行必需文件外，LICENSE/NOTICE/README 等许可与出处文件以及 preview/
-// 目录（皮肤预览图）一并随包分发。
-// V4：先比对「源 vs 目标」的内容戳记（版本+文件数+字节数），一致则跳过 ——
-// 旧逻辑每次启动全量重拷（dsh-pet 15MB、dsh-dafeiyu ~58MB 资产，拖慢启动）。
-// 戳记文件放在包目录内（.eac-copy-stamp.json），pnpm 重写 node_modules 时
-// 随目录消失，天然触发重建。
-const COPY_STAMP = '.eac-copy-stamp.json';
-
-// 与 copyPluginPackage 的拷贝清单保持一致（多算/漏算都会导致每次都重拷，
-// 只会浪费不会出错）。
-function pluginCopyEntries(src) {
-  const out = [];
-  const copyFile = (rel) => {
-    const sf = path.join(src, rel);
-    if (!fs.existsSync(sf) || fs.statSync(sf).isDirectory()) return;
-    out.push(rel);
-  };
-  const copyDir = (rel) => {
-    const sd = path.join(src, rel);
-    if (!fs.existsSync(sd) || !fs.statSync(sd).isDirectory()) return;
-    for (const entry of fs.readdirSync(sd, { withFileTypes: true })) {
-      const sub = rel + '/' + entry.name;
-      if (entry.isDirectory()) copyDir(sub);
-      else copyFile(sub);
-    }
-  };
-  for (const f of ['package.json', 'skin.json', ...EXTRA_PACKAGE_FILES]) copyFile(f);
-  for (const f of ['index.js', 'client.js', 'recall-inject.js', 'cordis.patch.yml']) copyFile(f);
-  for (const d of ['lib', 'preview', 'vendor', 'node_modules', 'data', 'assets', 'runtime', 'src', 'client']) copyDir(d);
-  return out;
-}
-
-function pluginStampOf(src) {
-  try {
-    const pkg = readJsonFile(path.join(src, 'package.json')) || {};
-    let files = 0;
-    let bytes = 0;
-    for (const rel of pluginCopyEntries(src)) {
-      files += 1;
-      try { bytes += fs.statSync(path.join(src, rel)).size; } catch {}
-    }
-    return JSON.stringify({ v: String(pkg.version || ''), f: files, b: bytes });
-  } catch {
-    return null;
-  }
-}
-
-function copyPluginPackage(profileDirP, src, name) {
-  const destRoot = path.join(profileDirP, 'node_modules', ...name.split('/'));
-  const stampFile = path.join(destRoot, COPY_STAMP);
-  const want = pluginStampOf(src);
-  try {
-    if (want && fs.existsSync(stampFile) && fs.readFileSync(stampFile, 'utf8') === want) {
-      return; // 内容未变：跳过全量重拷
-    }
-  } catch { /* 比对失败按需重拷 */ }
-  fs.mkdirSync(path.dirname(destRoot), { recursive: true });
-  const copyFile = (rel) => {
-    const sf = path.join(src, rel);
-    if (!fs.existsSync(sf) || fs.statSync(sf).isDirectory()) return;
-    const df = path.join(destRoot, rel);
-    fs.mkdirSync(path.dirname(df), { recursive: true });
-    fs.copyFileSync(sf, df);
-  };
-  const copyDir = (rel) => {
-    const sd = path.join(src, rel);
-    if (!fs.existsSync(sd) || !fs.statSync(sd).isDirectory()) return;
-    for (const entry of fs.readdirSync(sd, { withFileTypes: true })) {
-      const sub = rel + '/' + entry.name;
-      if (entry.isDirectory()) copyDir(sub);
-      else copyFile(sub);
-    }
-  };
-  // lib 整目录随包（配套插件可能有 logic.js 等额外模块，按清单拷会漏文件
-  // 导致 dsh web 启动时 ERR_MODULE_NOT_FOUND）。
-  for (const f of ['package.json', 'skin.json', ...EXTRA_PACKAGE_FILES]) copyFile(f);
-  // 社区插件（soul-md / tdai-memory / tool-vision）入口在包根目录而非
-  // lib/，vendor/ 是其内置依赖，同样必须随包分发。
-  for (const f of ['index.js', 'client.js', 'recall-inject.js', 'cordis.patch.yml']) copyFile(f);
-  copyDir('lib');
-  copyDir('preview');
-  copyDir('vendor');
-  // 内置插件自带的嵌套 node_modules（vendored 运行时依赖）：放在包内部，
-  // pnpm 重写 profile node_modules 顶层时不会波及，插件保持自包含。
-  copyDir('node_modules');
-  // dsh-webui-market 的离线目录快照（官网不可达时的兜底数据）。
-  copyDir('data');
-  // dsh-pet / dsh-dafeiyu 等带运行时静态资源的插件（宠物动画 webp/png 帧、
-  // PyInstaller helper 等）。
-  copyDir('assets');
-  copyDir('runtime');
-  // dsh-dafeiyu 的入口在 src/（lib/ 只有 client 半边）；dsh-offpeak 的
-  // client 半边在 client/（包 exports 映射）。
-  copyDir('src');
-  copyDir('client');
-  if (want) {
-    try {
-      fs.mkdirSync(destRoot, { recursive: true });
-      fs.writeFileSync(stampFile, want);
-    } catch { /* 戳记写失败不影响功能 */ }
-  }
-}
-
-// 随插件/皮肤包一起拷贝到 profile 的许可与出处文件（存在才拷贝）。
-const EXTRA_PACKAGE_FILES = ['LICENSE', 'LICENSE.md', 'NOTICE', 'NOTICE.md', 'README.md', 'README.zh.md', 'THIRD-PARTY-NOTICES.md'];
-
-// pnpm（dsh plugin add / 插件市场）hoist 进 profile node_modules 的
-// @deepseek-ai 核心包真实拷贝，会遮蔽 <home>/profiles/node_modules 里指向
-// 随应用分发的安装闭包 junction，形成模块双实例：Symbol 身份不一致，
-// 作用域注册失效（如 "deployment:persona is already registered"），
-// 模型列表刷新、模式切换、工作区添加等全部瘫痪。启动时清掉这些
-// 遮蔽拷贝，让解析回落到 junction —— 与宿主同源、全局单实例。
-function healProfileModules() {
-  try {
-    const home = state.dshHome || path.join(os.homedir(), '.dsh');
-    const removed = healProfileModuleShadowing(home, desktopProfile());
-    if (removed.length) log('boot', '已清理 profile node_modules 中遮蔽安装闭包的包拷贝: ' + removed.join(', '));
-  } catch (err) {
-    log('boot', '清理 profile 模块遮蔽失败: ' + err.message);
-  }
-}
-
-// ---------------------------------------------------------------------------
-// 插件市场排队任务：服务运行中安装/卸载撞上 Windows 文件锁（EPERM，如
-// sqlite-vec 的 vec0.dll 被运行中的 web 进程加载）时，市场插件把任务写进
-// profile 的 .dsh-market-pending.json。这里在"无服务进程持锁"的窗口期
-// （应用启动时 / 原地重启 kill 完旧进程后）用 dsh CLI 完成它。
-// ---------------------------------------------------------------------------
-const MARKER_NAME = '.dsh-market-pending.json';
-const MARKER_MAX_ATTEMPTS = 3;
-
-// 删除排队标记文件。曾有残留进程短暂持锁导致 rmSync 静默失败、标记
-// "复活"并反复触发 pnpm 的案例 —— 这里带重试 + 改名兜底，并返回是否
-// 真正删除，调用方据此决定是否放弃任务。
-function removeMarkerFile(file) {
-  try {
-    fs.rmSync(file, { force: true, maxRetries: 5, retryDelay: 200 });
-  } catch { /* 落到改名兜底 */ }
-  if (!fs.existsSync(file)) return true;
-  try {
-    fs.renameSync(file, file + '.stale-' + Date.now());
-  } catch { /* 锁着也无可奈何，交给 attempts 上限 */ }
-  return !fs.existsSync(file);
-}
-
-function pendingMarketMarkers() {
-  const out = [];
-  try {
-    const home = state.dshHome || path.join(os.homedir(), '.dsh');
-    const profilesRoot = path.join(home, 'profiles');
-    if (!fs.existsSync(profilesRoot)) return out;
-    for (const entry of fs.readdirSync(profilesRoot, { withFileTypes: true })) {
-      if (!entry.isDirectory()) continue;
-      const marker = path.join(profilesRoot, entry.name, MARKER_NAME);
-      if (!fs.existsSync(marker)) continue;
-      try {
-        // 去掉可能的 UTF-8 BOM（外部编辑器写入的标记）再解析。
-        const job = JSON.parse(fs.readFileSync(marker, 'utf8').replace(/^\uFEFF/, ''));
-        if (job && typeof job.target === 'string' && job.target
-          && typeof job.profile === 'string' && /^[A-Za-z0-9_-]+$/.test(job.profile)
-          && (job.kind === 'install' || job.kind === 'uninstall')) {
-          // V4.2：旧版 host 可能把目录默认 profile 'web' 写进标记（桌面壳跑
-          // 在 web-desktop，profiles/web 不存在）—— 归一化后再执行，避免对
-          // 不存在的 profile 跑 pnpm（spawn 报 node.exe ENOENT）。
-          job.profile = job.profile === 'web' ? desktopProfile() : job.profile;
-          out.push({ marker, job });
-        } else {
-          log('market-pending', '标记字段不完整，已删除: ' + marker);
-          removeMarkerFile(marker);
-        }
-      } catch (err) {
-        log('market-pending', `标记损坏，已删除: ${marker} (${err.message})`);
-        removeMarkerFile(marker);
-      }
-    }
-  } catch (err) {
-    log('market-pending', '扫描排队任务失败: ' + err.message);
-  }
-  return out;
-}
-
-function finishMarketMarker(marker, job, attempts, ok, tail) {
-  if (ok) {
-    log('market-pending', '排队任务完成: ' + (job.label || job.target));
-    if (!removeMarkerFile(marker)) {
-      log('market-pending', '警告: 排队标记删除失败（文件被占用？），已尝试改名兜底');
-    }
-    return;
-  }
-  if (attempts >= MARKER_MAX_ATTEMPTS) {
-    const last = String(tail || '').split(/\r?\n/).filter(Boolean).pop() || '';
-    log('market-pending', `排队任务连续 ${attempts} 次失败，放弃并清除: ${job.label || job.target}${last ? ' — ' + last.slice(0, 200) : ''}`);
-    removeMarkerFile(marker);
-    return;
-  }
-  try { fs.writeFileSync(marker, JSON.stringify({ ...job, attempts }, null, 2)); } catch {}
-  log('market-pending', '排队任务失败（下次启动重试）: ' + (job.label || job.target));
-}
 
 
-
-// 由桌面壳重建的包（配套插件 + 皮肤）不进快照：丢了也会被 syncCompanion
-// Plugins / 皮肤同步立刻补回，缓存它们只浪费空间。
-function managedPackageNames() {
-  const names = COMPANION_PLUGINS.map((p) => p.name);
-  try {
-    for (const entry of fs.readdirSync(SKINS_DIR, { withFileTypes: true })) {
-      if (!entry.isDirectory()) continue;
-      const pkg = readJsonFile(path.join(SKINS_DIR, entry.name, 'package.json'));
-      if (pkg && typeof pkg.name === 'string') names.push(pkg.name);
-    }
-  } catch {}
-  return names;
-}
-
-// 启动兜底回填：上次 pnpm 运行后若应用异常退出没来得及回填（或回填被
-// 中断），这里补上。只补缺失文件，安全幂等。
-async function restoreKeptArtifacts(profile) {
-  const ak = await artifactKeep();
-  if (typeof ak.restoreArtifacts !== 'function') return;
-  try {
-    ak.restoreArtifacts(profileDirFor(profile), artifactCacheDirFor(profile), {
-      log: (m) => log('artifact-keep', m),
-    });
-  } catch (err) {
-    log('artifact-keep', '回填失败: ' + err.message);
-  }
-}
-
-// 必须在"没有任何 dsh web 进程持锁"时调用；调用方负责先等待旧进程退出。
-async function processPendingMarketOps() {
-  const items = pendingMarketMarkers();
-  if (items.length === 0) return;
-  const nodeBin = nodeExe();
-  const bin = dshBin();
-  if (!fs.existsSync(nodeBin) || !fs.existsSync(bin)) {
-    log('market-pending', '找不到 node/dsh CLI，跳过排队任务');
-    return;
-  }
-  log('market-pending', `发现 ${items.length} 个排队任务，开始执行（Web 服务启动前，无文件锁）`);
-  // V4：pnpm 即将重写 node_modules —— 先快照第三方包（含人工补齐的
-  // lib/ 等构建产物），任务结束后回填被清掉的部分（meow-memory 修复）。
-  const profiles = [...new Set(items.map((it) => it.job.profile))];
-  const ak = await artifactKeep();
-  if (typeof ak.snapshotArtifacts === 'function') {
-    for (const profile of profiles) {
-      try {
-        ak.snapshotArtifacts(profileDirFor(profile), artifactCacheDirFor(profile), {
-          managedNames: managedPackageNames(),
-          log: (m) => log('artifact-keep', m),
-        });
-      } catch (err) {
-        log('artifact-keep', `snapshot ${profile} 失败: ` + err.message);
-      }
-    }
-  }
-  await new Promise((resolve) => {
-    let idx = 0;
-    // V4.2：allowBuilds 自动放行后的重试只允许一次（同一 marker）。
-    const retriedMarkers = new Set();
-    const next = async () => {
-      if (idx >= items.length) {
-        // pnpm 可能重新 hoist 出 @deepseek-ai 遮蔽拷贝，装完立刻清理，
-        // 避免模块双实例（Symbol 身份不一致）问题拖到下次启动。
-        healProfileModules();
-        return resolve();
-      }
-      const { marker, job } = items[idx];
-      const retried = retriedMarkers.has(marker);
-      const attempts = Number(job.attempts || 0) + 1;
-      const action = job.kind === 'uninstall' ? 'remove' : 'add';
-      // 安装前快照（保护中心）：排队任务改的是 profile 配置面，出问题可
-      // 一键/自动回滚到这里。
-      ensureGuard().snapshot('market:' + job.target);
-      log('market-pending', `执行(${attempts}/${MARKER_MAX_ATTEMPTS}): dsh plugin --profile ${job.profile} ${action} ${job.target}`);
-      const child = spawn(nodeBin, [bin, 'plugin', '--profile', job.profile, action, job.target], {
-        cwd: state.userDataDir,
-        // CI=true 与市场插件 host 侧一致：pnpm v10 无 TTY 时对被忽略的构建
-        // 脚本（如 node-llama-cpp）静默放行，而不是 ERR_PNPM_IGNORED_BUILDS 硬失败。
-        env: { ...childEnv(), CI: 'true' },
-        windowsHide: true, stdio: ['ignore', 'pipe', 'pipe'],
-      });
-      state.marketOpChild = child;
-      let tail = '';
-      const onData = (c) => {
-        const text = c.toString();
-        tail = (tail + text).slice(-8000);
-        for (const line of text.split(/\r?\n/)) {
-          const s = line.trim();
-          // Progress: \r 进度条不进日志，只保留有信息量的行。
-          if (s && !/^Progress:/.test(s)) log('market-pending', s.slice(0, 300));
-        }
-      };
-      child.stdout.on('data', onData);
-      child.stderr.on('data', onData);
-      const timer = setTimeout(() => {
-        log('market-pending', '排队任务超时（5 分钟），强制终止');
-        try { spawn('taskkill', ['/pid', String(child.pid), '/T', '/F'], { windowsHide: true, stdio: 'ignore' }); } catch {}
-      }, 5 * 60 * 1000);
-      child.on('error', (err) => {
-        clearTimeout(timer);
-        if (state.marketOpChild === child) state.marketOpChild = null;
-        finishMarketMarker(marker, job, attempts, false, String(err.message));
-        idx += 1;
-        next();
-      });
-      child.on('close', async (code) => {
-        clearTimeout(timer);
-        if (state.marketOpChild === child) state.marketOpChild = null;
-        // V4.2：pnpm 封锁构建脚本硬失败时，从输出解析包名、自动写入
-        // pnpm-workspace.yaml 的 allowBuilds（兼容旧名 onlyBuiltDependencies）
-        // 后重试同一任务一次（不消耗 attempts）。
-        if (code !== 0 && !retried) {
-          try {
-            const ab = await allowBuilds();
-            const keys = (ab.parseBlockedBuildKeys || (() => []))(tail);
-            if (keys.length > 0) {
-              const r = await ab.ensureAllowBuilds(path.join(profileDirFor(job.profile), 'pnpm-workspace.yaml'), keys);
-              if (r && r.wrote) {
-                log('market-pending', `[allowBuilds] 已自动放行 ${r.added.join(', ')}，自动重试`);
-                retriedMarkers.add(marker);
-                next();
-                return;
-              }
-            }
-          } catch (err) {
-            log('market-pending', '[allowBuilds] 自动放行失败: ' + String((err && err.message) || err));
-          }
-        }
-        finishMarketMarker(marker, job, attempts, code === 0, tail);
-        idx += 1;
-        next();
-      });
-    };
-    next();
-  });
-  // pnpm 重写完成：回填被清掉的第三方构建产物（lib/ 等）。
-  if (typeof ak.restoreArtifacts === 'function') {
-    for (const profile of profiles) {
-      try {
-        ak.restoreArtifacts(profileDirFor(profile), artifactCacheDirFor(profile), {
-          log: (m) => log('artifact-keep', m),
-        });
-      } catch (err) {
-        log('artifact-keep', `restore ${profile} 失败: ` + err.message);
-      }
-    }
-  }
-}
-
-// 内置 skills 分发目录：assets/skills/<kebab-name>/SKILL.md。~/.dsh/skills
-// 本就是 dsh-skill-filesystem 的默认扫描根（rank 400），这里只需把内置
-// 技能同步过去 —— 内核零配置。同步规则：带 .eac-skill.json 标记的目录由
-// EAC 管理（版本变化时覆盖更新）；用户自建同名目录（无标记）永不覆盖。
-const BUNDLED_SKILLS_DIR = path.join(__dirname, 'assets', 'skills');
-
-function syncBundledSkills() {
-  try {
-    const src = BUNDLED_SKILLS_DIR;
-    if (!fs.existsSync(src)) return;
-    const destRoot = path.join(state.dshHome || path.join(os.homedir(), '.dsh'), 'skills');
-    fs.mkdirSync(destRoot, { recursive: true });
-    const installed = [];
-    for (const entry of fs.readdirSync(src, { withFileTypes: true })) {
-      if (!entry.isDirectory()) continue;
-      const skillSrc = path.join(src, entry.name);
-      if (!fs.existsSync(path.join(skillSrc, 'SKILL.md'))) continue;
-      const skillDst = path.join(destRoot, entry.name);
-      const markerSrc = readJsonFile(path.join(skillSrc, '.eac-skill.json')) || { version: 1, managed: true };
-      const markerDst = readJsonFile(path.join(skillDst, '.eac-skill.json'));
-      if (markerDst && markerDst.version === markerSrc.version) continue;
-      if (!markerDst && fs.existsSync(skillDst)) continue; // 用户自建同名技能：不动
-      fs.cpSync(skillSrc, skillDst, { recursive: true });
-      installed.push(entry.name);
-    }
-    if (installed.length) log('boot', '已同步内置 skills 到 ' + destRoot + ': ' + installed.join(', '));
-  } catch (err) {
-    log('boot', '同步内置 skills 失败: ' + err.message);
-  }
-}
-
-// ---------------------------------------------------------------------------
-// V4 运行时补丁（移植自上游 dsh_desktop，幂等）：覆盖三处运行副本 ——
-// profile 共享 junction 根、内置 app 副本（__dirname/node_modules）、用户
-// 更新过的 agent overlay（<userData>/agent/node_modules）。agent 更新会换掉
-// overlay 整树，补丁随 syncCompanionPlugins 每次启动重放。
-// ---------------------------------------------------------------------------
-
-function runtimePatchRoots() {
-  const home = state.dshHome || path.join(os.homedir(), '.dsh');
-  return [
-    path.join(home, 'profiles', 'node_modules'),
-    path.join(__dirname, 'node_modules'),
-    path.join(state.userDataDir, 'agent', 'node_modules'),
-  ];
-}
-
-// 对话删除 / 归档管理（dsh-session-manager 插件的前置依赖）：
-// dsh-workspace + dsh-host-apiproxy + dsh-session + dsh-client-connection +
-// dsh-client-ui-workspace 的外科手术式扩展（详见 scripts/patch-session-manage.js
-// 头注释）。锚点不匹配（官方包结构变化）时自动跳过，绝不损坏文件。
-function applySessionManageFix() {
-  for (const root of runtimePatchRoots()) {
-    if (!root || !fs.existsSync(root)) continue;
-    try {
-      const n = patchSessionManage(root, (m) => log('boot', m));
-      if (n > 0) log('boot', '对话删除补丁: 已应用到 ' + root);
-    } catch (err) {
-      log('boot', '对话删除补丁失败(' + root + '): ' + err.message);
-    }
-  }
-}
-
-// ---------------------------------------------------------------------------
-// 插件启停管理（V4，移植自上游）：设置页「插件 → 管理」标签的数据与写盘。
-// dsh:plugin-list / dsh:plugin-set-enabled 两个 IPC 驱动；写盘用纯文本手术
-// （scripts/plugin-manager-patch.js），保留文件其它内容与注释。
-// ---------------------------------------------------------------------------
-
-// 惰性加载 js-yaml（内置 dsh 的传递依赖）；缺失时管理页降级为空列表。
-function loadDshYamlDialect() {
-  if (state.dshYamlTried) return state.dshYamlDialect;
-  state.dshYamlTried = true;
-  try {
-    const yaml = require('js-yaml');
-    // 与 dsh 相同的 entry-list 方言：`!!js` 表达式是合法标量。
-    const jsType = new yaml.Type('tag:yaml.org,2002:js', {
-      kind: 'scalar',
-      resolve: (data) => typeof data === 'string',
-      construct: (data) => ({ __jsExpr: data }),
-    });
-    state.dshYamlDialect = { load: (content) => yaml.load(content, { schema: yaml.JSON_SCHEMA.extend(jsType) }) };
-  } catch {
-    state.dshYamlDialect = null;
-  }
-  return state.dshYamlDialect;
-}
-
-function pluginManagerReadPatch() {
-  const file = path.join(desktopProfileDir(), 'cordis.patch.yml');
-  let text = '';
-  try { text = fs.readFileSync(file, 'utf8'); } catch {}
-  const yaml = loadDshYamlDialect();
-  if (!yaml) return { file, text, entries: [] };
-  try {
-    const parsed = yaml.load(text);
-    return { file, text, entries: Array.isArray(parsed) ? parsed : [] };
-  } catch {
-    return { file, text, entries: [] };
-  }
-}
-
-function pluginManagerPackageDescription(name) {
-  if (!name) return '';
-  const candidates = [
-    path.join(desktopProfileDir(), 'node_modules', ...name.split('/')),
-    path.join(__dirname, 'assets', 'plugins', name.includes('/') ? name.slice(name.indexOf('/') + 1) : name),
-  ];
-  for (const dir of candidates) {
-    try {
-      const pkg = JSON.parse(fs.readFileSync(path.join(dir, 'package.json'), 'utf8'));
-      if (pkg && typeof pkg.description === 'string' && pkg.description) return pkg.description;
-    } catch {}
-  }
-  return '';
-}
-
-function pluginManagerCollect() {
-  const { entries } = pluginManagerReadPatch();
-  let bundles = [];
-  try {
-    const m = JSON.parse(fs.readFileSync(path.join(desktopProfileDir(), 'package.json'), 'utf8'));
-    bundles = (m && m.dsh && m.dsh.profile && Array.isArray(m.dsh.profile.bundles)) ? m.dsh.profile.bundles : [];
-  } catch {}
-  return collectPluginRows(entries, {
-    companion: COMPANION_PLUGINS.map((p) => ({ id: p.id, name: p.name })),
-    coreIds: onboardingLogic.CORE_PLUGIN_IDS,
-    removedIds: removedPluginIds(),
-    describe: (name) => pluginManagerPackageDescription(name),
-    bundles,
-  });
-}
-
-function pluginManagerResolveName(id) {
-  const c = COMPANION_PLUGINS.find((p) => p.id === id);
-  if (c) return c.name;
-  const { entries } = pluginManagerReadPatch();
-  for (const entry of entries) {
-    if (entry && Array.isArray(entry.insert)) {
-      const it = entry.insert.find((x) => x && x.id === id);
-      if (it && it.name) return it.name;
-    }
-  }
-  return '';
-}
-
-// ---------------------------------------------------------------------------
-// 内置插件「移除」（V4.2）：把插件的 id 记入 settings.removedPlugins 跳过
-// syncCompanionPlugins，同时清掉 profile 里的 patch 行与 node_modules 副本。
-// 区别于「禁用」（停用但保留，随时可开）——移除是卸载语义，重启不还原；
-// 市场里重复安装同名内置包也不被拒绝（内置清单已不含它）。
-// ---------------------------------------------------------------------------
-
-function removedPluginIds() {
-  try {
-    const s = updater.loadSettings(updCtx());
-    return new Set(Array.isArray(s.removedPlugins) ? s.removedPlugins : []);
-  } catch { return new Set(); }
-}
-
-function saveRemovedPluginIds(ids) {
-  const ctx = updCtx();
-  const s = updater.loadSettings(ctx);
-  s.removedPlugins = Array.from(ids);
-  updater.saveSettings(ctx, s);
-}
-
-// 恢复单个配套插件：立即复制包 + 补写 patch 行（与 syncCompanionPlugins
-// 的写入规则一致），重启服务后生效。源目录走「覆盖层优先」（V4.3）：
-// 被恢复的内置插件若是已更新版本，恢复回来的就是更新版。
-function restoreCompanionPlugin(p) {
-  const profileDirP = desktopProfileDir();
-  const dirName = p.dir || (p.name.includes('/') ? p.name.split('/').pop() : p.name);
-  const src = builtinPluginSourceDir(dirName);
-  if (!fs.existsSync(path.join(src, 'package.json'))) {
-    return { ok: false, error: '配套插件源目录无效: ' + src };
-  }
-  copyPluginPackage(profileDirP, src, p.name);
-  const patchFile = path.join(profileDirP, 'cordis.patch.yml');
-  let patch = '';
-  try { patch = fs.readFileSync(patchFile, 'utf8'); } catch {}
-  if (!hasEntryId(patch, p.id)) {
-    let bundled = [];
-    try { bundled = readJsonFile(path.join(profileDirP, 'package.json'))?.dsh?.profile?.bundles || []; } catch { bundled = []; }
-    if (!bundled.includes(p.name)) {
-      let block = `- insert:\n    - id: ${p.id}\n      name: '${p.name}'\n`;
-      if (p.config) block += configLinesFor(p.config);
-      if (p.disabled) block += `      disabled: true\n`;
-      if (/^\s*\[\]\s*$/m.test(patch)) patch = patch.replace(/\[\]/m, block);
-      else if (patch.trim() === '') patch = '# dsh web profile patch（由 DSH Desktop 维护）\n' + block;
-      else patch = patch.replace(/\s*$/, '\n') + block;
-      try { fs.writeFileSync(patchFile, patch); } catch (err) {
-        return { ok: false, error: String((err && err.message) || err) };
-      }
-    }
-  }
-  return { ok: true };
-}
-
-// removed=true 移除（卸载语义）；removed=false 恢复。核心插件拒绝移除。
-function pluginManagerSetRemoved(id, removed) {
-  const p = COMPANION_PLUGINS.find((x) => x.id === id);
-  if (!p) return { ok: false, error: '未知内置插件: ' + String(id) };
-  if (onboardingLogic.CORE_PLUGIN_IDS.has(id)) {
-    return { ok: false, error: '核心插件不可移除: ' + String(id) };
-  }
-  const removedSet = removedPluginIds();
-  const patchFile = path.join(desktopProfileDir(), 'cordis.patch.yml');
-  try {
-    if (removed) {
-      // 1) 清 patch 行（顶层 + insert 内层）
-      let text = '';
-      try { text = fs.readFileSync(patchFile, 'utf8'); } catch {}
-      const patched = removePluginFromPatch(text, id);
-      if (patched !== text) fs.writeFileSync(patchFile, patched, 'utf8');
-      // 2) 删 profile node_modules 里的包副本（copyPluginPackage 的产物）
-      const pkgDir = path.join(desktopProfileDir(), 'node_modules', p.name);
-      fs.rmSync(pkgDir, { recursive: true, force: true });
-      // 3) 记入跳过清单（下次 sync 不再写回）
-      removedSet.add(id);
-      saveRemovedPluginIds(removedSet);
-      log('plugin-manager', '已移除内置插件 ' + id);
-      return { ok: true, restartRequired: true };
-    }
-    // 恢复：清出跳过清单 + 立即复制包与行
-    removedSet.delete(id);
-    saveRemovedPluginIds(removedSet);
-    const res = restoreCompanionPlugin(p);
-    if (!res.ok) return res;
-    log('plugin-manager', '已恢复内置插件 ' + id);
-    return { ok: true, restartRequired: true };
-  } catch (err) {
-    log('plugin-manager', '移除/恢复插件 ' + id + ' 失败: ' + ((err && err.message) || err));
-    return { ok: false, error: String((err && err.message) || err) };
-  }
-}
-
-// 图片粘贴保存（dsh-image-paste 插件）：只接受 image/* 的 data URL，
-// base64 解码后原子写入 %TEMP%/dsh-paste/<清洗名>-<时间戳><ext>，返回
-// { ok, path, size }。文件在临时目录，随系统清理，不污染工作区。
-const IMAGE_PASTE_MAX_BYTES = 15 * 1024 * 1024;
-const IMAGE_PASTE_EXT = {
-  'image/png': '.png',
-  'image/jpeg': '.jpg',
-  'image/webp': '.webp',
-  'image/gif': '.gif',
-  'image/bmp': '.bmp',
-  'image/avif': '.avif',
-  'image/ico': '.ico',
-  'image/x-icon': '.ico',
-  'image/tiff': '.tiff',
-};
-
-function imagePasteSave(dataUrl, name) {
-  const m = /^data:(image\/[\w.+-]+);base64,([A-Za-z0-9+/=]+)$/.exec(dataUrl);
-  if (!m) return { ok: false, error: '不是合法的图片 data URL' };
-  const mime = m[1].toLowerCase();
-  if (!IMAGE_PASTE_EXT[mime]) return { ok: false, error: '不支持的图片类型: ' + mime };
-  const buf = Buffer.from(m[2], 'base64');
-  if (buf.length === 0) return { ok: false, error: '图片内容为空' };
-  if (buf.length > IMAGE_PASTE_MAX_BYTES) return { ok: false, error: '图片超过 15MB 上限' };
-  const dir = path.join(os.tmpdir(), 'dsh-paste');
-  fs.mkdirSync(dir, { recursive: true });
-  const base = String(name).replace(/[\\/:*?"<>|\u0000-\u001f]/g, '_').trim().slice(0, 40) || '粘贴图片';
-  const file = path.join(dir, base + '-' + Date.now() + IMAGE_PASTE_EXT[mime]);
-  fs.writeFileSync(file, buf);
-  return { ok: true, path: file, size: buf.length };
-}
-
-// 写入/移除用户层 disabled 条目（纯文本手术见 scripts/plugin-manager-patch.js）：
-// 与上游的差异 —— 「启用」保留顶层裸条目 {id, name} 而不是整条移除，这样
-// 默认禁用的配套插件（dsh-pet）被用户启用后不会被下次 sync 重新插回
-// disabled 行（sync 的「已有行不重写」规则自然接管）。
-function pluginManagerSetEnabled(id, enabled) {
-  const file = path.join(desktopProfileDir(), 'cordis.patch.yml');
-  let text = '';
-  try { text = fs.readFileSync(file, 'utf8'); } catch {}
-  if (!text.trim()) text = '# dsh web profile patch（由 Deepseek Harness EAC 维护）\n';
-
-  const name = pluginManagerResolveName(id);
-  if (!enabled && !name) return { ok: false, error: '无法解析插件包名: ' + id };
-
-  let patched;
-  try {
-    patched = togglePluginInPatch(text, id, !!enabled, name);
-  } catch (err) {
-    return { ok: false, error: String((err && err.message) || err) };
-  }
-  if (patched !== text) {
-    try {
-      const tmp = file + '.tmp';
-      fs.writeFileSync(tmp, patched, 'utf8');
-      fs.renameSync(tmp, file);
-    } catch (err) {
-      return { ok: false, error: String((err && err.message) || err) };
-    }
-  }
-  return { ok: true };
-}
-
-function syncCompanionPlugins() {
-  if (!IS_WIN) return;
-  try {
-    const home = state.dshHome || path.join(os.homedir(), '.dsh');
-    // 桌面专属 profile 必须先存在（未知 profile 不会被 dsh 自动初始化）。
-    ensureDesktopProfileInit();
-    // V4 运行时补丁（幂等，随启动 / 服务重启 / agent 更新后重放）：
-    //  · 对话删除/归档 —— dsh-session-manager 插件的全链路前置依赖；
-    applySessionManageFix();
-    const profileDirP = desktopProfileDir();
-    // 内置社区 agent preset（anchored-standard：首请求锚定 Minimal 工具对，
-    // 首次工具调用/回复后开放完整 Standard 目录）：安装到用户 preset 根。
-    // preset 不进插件树，坏 preset 不会拖垮启动；已存在则跳过（用户手装
-    // 或改过的版本优先），见 preset-sync.js。
-    const presetsSynced = syncBundledPresets(
-      path.join(__dirname, 'assets', 'agent-presets'),
-      path.join(home, '.agent-presets'),
-      (m) => log('boot', m)
-    );
-    if (presetsSynced.installed.length) log('boot', '已安装内置 agent preset: ' + presetsSynced.installed.join(', '));
-    // 默认 preset 指到内置的 anchored-standard（用户已在 settings.yaml 写过
-    // default 则一律保留）。失败只降级为官方默认 preset，不影响启动。
-    const defaultResult = ensureDefaultAgentPreset(home, 'anchored-standard', (m) => log('boot', m));
-    if (defaultResult === 'set') log('boot', '已设置默认 agent preset: anchored-standard');
-    else if (defaultResult === 'kept') log('boot', '用户已设置默认 agent preset，保持不变');
-    fs.mkdirSync(path.join(profileDirP, 'node_modules'), { recursive: true });
-    const pending = [];
-    const removedIds = removedPluginIds();
-    // V4.2：用户曾从市场安装过与内置插件同名的包时，写包前先迁移残留
-    // （package.json 依赖/bundles + patch 行），让内置版干净接管，避免
-    // duplicate loader entry；完成后系统通知告知「插件树变化」。
-    const migratedBuiltins = [];
-    for (const p of COMPANION_PLUGINS) {
-      // V4.2：用户移除过的内置插件不再复制/登记（见 pluginManagerSetRemoved）。
-      if (removedIds.has(p.id)) {
-        log('boot', `已按用户选择跳过被移除的内置插件: ${p.id}`);
-        continue;
-      }
-      // 非 @deepseek-ai 作用域的配套包用显式 dir 指定 assets/plugins 下的目录名；
-      // 回退解析按「最后一个路径段」取（@scope/name → name；无 scope → 原名）。
-      // V4 修复：旧回退是 name.slice('@deepseek-ai/'.length) —— 对无 scope 的
-      // 长包名会截出错误目录（dsh-session-manager → 'manager'），该插件被
-      // 静默跳过（行与包都不落盘）。
-      const dirName = p.dir || (p.name.includes('/') ? p.name.split('/').pop() : p.name);
-      // V4.3：覆盖层优先 —— 用户更新过的内置插件从 <userData>/builtin-plugin-updates
-      // 拷贝（不被资产版本还原）；应用升级后资产版本更新则自动接管。
-      const src = builtinPluginSourceDir(dirName);
-      if (!fs.existsSync(path.join(src, 'package.json'))) {
-        log('boot', `配套插件源目录无效，跳过: ${p.id} → ${src}`);
-        continue;
-      }
-      try {
-        const { removeMarketDuplicate, patchHasForeignRows } = require('./builtin-collision');
-        // 市场同名包残留预检（v4.2，用户反馈问题 5）：只有「非应用自写」证据
-        // （package.json 依赖/bundles 或非自写 patch 行）才算残留。
-        const dupPreCheck = (() => {
-          try {
-            const pkg = readJsonFile(path.join(profileDirP, 'package.json'));
-            const spec = pkg && pkg.dependencies && pkg.dependencies[p.name];
-            if (spec && !String(spec).startsWith('link:') && !String(spec).startsWith('file:')) return true;
-            if (pkg && pkg.dsh && pkg.dsh.profile && Array.isArray(pkg.dsh.profile.bundles) && pkg.dsh.profile.bundles.includes(p.name)) return true;
-            const patchText = fs.readFileSync(path.join(profileDirP, 'cordis.patch.yml'), 'utf8');
-            // 只认「非应用自写」的登记行：sync 的 insert 内层行、插件管理/向导
-            // togglePluginInPatch 写的（带「关闭」标记注释的）顶层行都是应用自己
-            // 的启停状态，不是市场残留。否则 v4.4 首次向导的取消勾选会在同一启动
-            // 里被剥离后按注册表默认回写（dsh-dafeiyu 等默认启用插件被静默重新
-            // 启用），且每次启动产生「剥离-回写」空转与孤儿 `- insert:` 行堆积。
-            return patchHasForeignRows(patchText, p.name);
-          } catch { return false; }
-        })();
-        if (dupPreCheck) {
-          // 先快照（保护中心）：迁移属于配置面手术，出问题可一键回滚。
-          ensureGuard().snapshot('builtin-migrate:' + p.id);
-          // 只在确有市场残留证据时才动手术。v4.2 曾无条件执行迁移 —— 它的
-          // 「剥离-回写」对无重复用户是空转，且会把应用自写的行（向导/插件
-          // 管理的 disabled 行、sync 自己的 insert 行）一并剥掉后按注册表
-          // 默认回写：v4.4 首次向导的取消勾选被静默重新启用，孤儿 insert
-          // 行每次启动堆积。
-          const migrated = removeMarketDuplicate(profileDirP, p.name, { log: (m) => log('boot', m) });
-          if (migrated.changed && migrated.ok) {
-            migratedBuiltins.push({ name: p.name, dep: migrated.removedDep.length > 0, rows: migrated.removedRows });
-            log('boot', `内置插件 ${p.name} 已接管市场同名包（移除依赖 ${migrated.removedDep.length} 个、patch 行 ${migrated.removedRows.length} 个）`);
-          }
-        }
-      } catch (err) {
-        log('boot', `内置插件同名迁移失败(${p.id}): ${String((err && err.message) || err)}`);
-      }
-      copyPluginPackage(profileDirP, src, p.name);
-      // p.disabled: true 的配套插件默认以禁用行注册（如 dsh-pet 页面桌宠），
-      // 用户可在「设置 → 插件 → 管理」里启用；已有行不重写，用户选择优先。
-      pending.push({ id: p.id, name: p.name, disabled: p.disabled === true, config: p.config });
-    }
-    if (migratedBuiltins.length) {
-      try {
-        const names = migratedBuiltins.map((m) => m.name).join('、');
-        const n = new Notification({
-          title: '内置插件已接管同名市场包',
-          body: `检测到市场安装的重复包，已改用内置版本（${names}）。插件树已自动整理，本次启动生效。`,
-          icon: path.join(__dirname, 'assets', 'icon.png'),
-        });
-        n.on('click', () => showMainWindow());
-        n.show();
-      } catch (err) {
-        log('boot', '内置接管通知发送失败: ' + err.message);
-      }
-    }
-    // 内置皮肤：行 id 取皮肤包 skin.json 的 wiring.id（ui-skin-*）。
-    for (const entry of fs.readdirSync(SKINS_DIR, { withFileTypes: true })) {
-      if (!entry.isDirectory()) continue;
-      const src = path.join(SKINS_DIR, entry.name);
-      const pkg = readJsonFile(path.join(src, 'package.json'));
-      if (!pkg || typeof pkg.name !== 'string' || !pkg.name.includes('/')) continue;
-      const skin = readJsonFile(path.join(src, 'skin.json'));
-      const rowId = skin && skin.wiring && typeof skin.wiring.id === 'string' ? skin.wiring.id : '';
-      if (!/^ui-skin-[\w-]+$/.test(rowId)) continue;
-      copyPluginPackage(profileDirP, src, pkg.name);
-      pending.push({ id: rowId, name: pkg.name, disabled: true });
-    }
-    // 内置插件清单标记：插件市场据此把目录里的同名插件标为「已内置」并
-    // 拒绝重复安装 —— 内置包每次启动都被重新同步，市场覆盖安装会产生
-    // duplicate loader entry / 模块双实例，必须从源头拦截。
-    try {
-      const builtinNames = pending.map((p) => p.name);
-      const marker = path.join(profileDirP, '.dsh-builtin-plugins.json');
-      const prev = readJsonFile(marker);
-      const next = { names: builtinNames, updatedAt: new Date().toISOString() };
-      if (!prev || JSON.stringify(prev.names) !== JSON.stringify(next.names)) {
-        fs.writeFileSync(marker, JSON.stringify(next, null, 2) + '\n');
-      }
-    } catch (err) {
-      log('boot', '写入内置插件清单失败: ' + err.message);
-    }
-    // 注册到 profile 的 patch 层（幂等：已有行不重写，用户选择的皮肤/disabled 状态保留）。
-    const patchFile = path.join(profileDirP, 'cordis.patch.yml');
-    let patch = '';
-    try { patch = fs.readFileSync(patchFile, 'utf8'); } catch { patch = ''; }
-    let changed = false;
-    // 先修存量坏行：v2.0.0 写入的 soul-md 行缺 config.path（见 patch-row-heal.js
-    // 头注释），不修则升级用户仍会 “dsh web 启动失败 (退出码 1)”。
-    const healed = healSoulMdPatchRow(patch);
-    if (healed.healed.length) {
-      patch = healed.patch;
-      changed = true;
-      log('boot', '已修复 profile patch 中缺 config.path 的 soul-md 行');
-    }
-    // V4：修复 v3.1.0 及以前写出的「无 config 的 dsh-pet 行」（loader 传
-    // undefined → dsh-pet 读 config.fullRoot 崩 → 插件树整体加载失败）。
-    const healedPet = healRowConfig(patch, 'dsh-pet', { size: 260, position: 'bottom-right' });
-    if (healedPet.healed.length) {
-      patch = healedPet.patch;
-      changed = true;
-      log('boot', '已修复 profile patch 中缺 config 的 dsh-pet 行（v3 存量坏行）');
-    }
-    // 市场安装（dsh plugin add）会把插件登记进 package.json 的
-    // dsh.profile.bundles，加载时执行其包内 patch 挂载行；若 overlay 里
-    // 也有一行（syncCompanionPlugins 写的），整个插件树会以
-    // “duplicate loader entry id” 崩溃。清掉 overlay 重复行（包内行保留）。
-    let bundled = [];
-    try { bundled = readJsonFile(path.join(profileDirP, 'package.json'))?.dsh?.profile?.bundles || []; } catch { bundled = []; }
-    // 同一 entry id 被两处声明（bundle 的包内 patch + overlay 的配套行）会以
-    // “duplicate loader entry id” 拖垮整个插件树。旧逻辑只按「包名 ∈ bundles」
-    // 匹配，git/fork/link 安装的插件包名与配套行包名不符时永远删不掉（issue
-    // #16）。这里再解析每个 bundle 包实际声明的 entry id 集合：overlay 中 id
-    // 已被任一 bundle 声明（无论包名如何）即视为重复。
-    const declaredBundleIds = collectBundleEntryIds(bundled, path.join(profileDirP, 'node_modules'));
-    const rowIds = {};
-    for (const p of COMPANION_PLUGINS) rowIds[p.id] = p.name;
-    const deduped = removeBundledRowDuplicates(patch, rowIds, bundled, declaredBundleIds);
-    if (deduped.removed.length) {
-      patch = deduped.patch;
-      changed = true;
-      log('boot', '已移除与 bundle 登记重复的 patch 行: ' + deduped.removed.join(', '));
-    }
-    for (const p of pending) {
-      if (hasEntryId(patch, p.id)) continue;
-      // 已在 bundle 列表里的插件由其包内 patch 挂载，overlay 不能再写行
-      // （会 duplicate loader entry id，拖垮整个插件树）。issue #16：
-      // 补充按 entry id 判断 —— git/fork 插件包名不同但 id 相同同样要跳过，
-      // 否则每次启动把崩溃行写回，用户删掉也没用。
-      if (bundled.includes(p.name) || declaredBundleIds.has(p.id)) continue;
-      let block = `- insert:\n    - id: ${p.id}\n      name: '${p.name}'\n`;
-      if (p.config) block += configLinesFor(p.config);
-      if (p.disabled) block += `      disabled: true\n`;
-      if (/^\s*\[\]\s*$/m.test(patch)) patch = patch.replace(/\[\]/m, block);
-      else if (patch.trim() === '') patch = '# dsh web profile patch（由 DSH Desktop 维护）\n' + block;
-      else patch = patch.replace(/\s*$/, '\n') + block;
-      changed = true;
-    }
-    if (changed) {
-      // 顺带清理历史遗留的孤儿 `- insert:` 行（v4.2/4.3 每次启动「剥离-回写」
-      // 残留的空块；对 cordis 无效果，仅文件卫生）。强制一次写盘，之后幂等。
-      const lines = patch.split(/\r?\n/);
-      const cleaned = lines.filter((line, idx) => {
-        if (!/^[ \t]*- insert:\s*$/.test(line)) return true;
-        let k = idx + 1;
-        while (k < lines.length && lines[k].trim() === '') k += 1;
-        return k < lines.length && /^[ \t]+- /.test(lines[k]);
-      }).join('\n');
-      if (cleaned !== patch) {
-        patch = cleaned;
-        log('boot', '已清理 profile patch 中的孤儿 - insert: 行');
-      }
-      fs.writeFileSync(patchFile, patch);
-      log('boot', '已同步配套插件/皮肤到 web profile: ' + pending.map((p) => p.id).join(', '));
-    }
-    // 迁移带来的皮肤选择（migrateFromSharedWebProfile 记录）在此落位。
-    applyLegacySkinChoice();
-  } catch (err) {
-    log('boot', '同步配套插件失败: ' + err.message);
-  }
-}
 
 // ---------------------------------------------------------------------------
 // 快捷方式维护：修复「没有桌面快捷方式 / 快捷方式指向的文件消失」，
@@ -2514,148 +1474,6 @@ function maintainShortcuts() {
   }
 }
 
-function warnTempRun() {
-  if (!app.isPackaged || !IS_WIN || !process.env.PORTABLE_EXECUTABLE_DIR) return;
-  // E2E（scripts/e2e-v4.js）从临时目录跑便携版：告警弹窗会卡住无头验证。
-  if (process.env.DSH_DESKTOP_TEST_NO_SHORTCUTS === '1') return;
-  const dir = process.env.PORTABLE_EXECUTABLE_DIR.toLowerCase();
-  const tmp = os.tmpdir().toLowerCase();
-  if (dir === tmp || dir.startsWith(tmp + path.sep)) {
-    showBox({
-      type: 'warning',
-      title: '正在从临时目录运行',
-      message: '当前便携版位于系统临时目录。',
-      detail: '临时目录中的文件可能被系统自动清理，导致快捷方式失效或程序“消失”。\n建议把 Deepseek Harness EAC exe 移动到固定位置（如桌面或 D 盘）后再运行。',
-      buttons: ['知道了'],
-    });
-  }
-}
-
-// ---------------------------------------------------------------------------
-// 一次性迁移：桌面端从共享 web profile 切到专属 web-desktop profile。
-//
-// 只做三件事，全部幂等：
-//   1. 记住用户在旧 profile 里启用的皮肤（迁移后在专属 profile 里复活）；
-//   2. 清掉旧 web profile 里桌面端写入的配套插件行 + 拷贝的配套包 + 内置
-//      清单标记 —— 原生 CLI 从此加载干净的 web profile（冲突面消除）；
-//   3. 标记 settings.desktopProfileMigrated，永不重复执行。
-// 用户用市场装进旧 profile 的插件（package.json bundles）是原生端资产，
-// 一律不动；桌面端如需继续使用，重新从市场安装即可（有保护中心兜底）。
-function migrateFromSharedWebProfile() {
-  try {
-    const s = updater.loadSettings(updCtx());
-    if (s.desktopProfileMigrated) return;
-    s.desktopProfileMigrated = new Date().toISOString();
-    updater.saveSettings(updCtx(), s); // 先落标记：即使下面失败也不反复折腾
-    if (s.shareWebProfile === true) return; // 用户显式选择共享模式
-
-    const home = state.dshHome || path.join(os.homedir(), '.dsh');
-    const oldDir = path.join(home, 'profiles', 'web');
-    const marker = path.join(oldDir, '.dsh-builtin-plugins.json');
-    if (!fs.existsSync(marker)) return; // 旧版本从没在共享 profile 跑过桌面端
-    const builtinNames = readJsonFile(marker)?.names || [];
-
-    // 1) 提取用户启用的皮肤行 id。
-    let enabledSkin = null;
-    const patchFile = path.join(oldDir, 'cordis.patch.yml');
-    let oldPatch = '';
-    try { oldPatch = fs.readFileSync(patchFile, 'utf8'); } catch { oldPatch = ''; }
-    {
-      const lines = oldPatch.split(/\r?\n/);
-      for (let i = 0; i < lines.length; i++) {
-        const m = /^- id: (ui-skin-[\w-]+)\s*$/.exec(lines[i]);
-        if (!m) continue;
-        let disabled = false;
-        for (let j = i + 1; j < lines.length; j++) {
-          if (/^- /.test(lines[j])) break;
-          if (/^\s+disabled:\s*true/.test(lines[j])) disabled = true;
-        }
-        if (!disabled) enabledSkin = m[1];
-      }
-    }
-
-    // 2) 清理旧 profile 的桌面端痕迹。
-    const rowIdSet = new Set();
-    for (const p of COMPANION_PLUGINS) rowIdSet.add(p.id);
-    for (const id of extractPatchRowIds(oldPatch)) {
-      if (/^ui-skin-[\w-]+$/.test(id)) rowIdSet.add(id);
-    }
-    const cleaned = removePatchRowsById(oldPatch, rowIdSet);
-    if (cleaned.removed.length) fs.writeFileSync(patchFile, cleaned.patch);
-    for (const name of builtinNames) {
-      try { fs.rmSync(path.join(oldDir, 'node_modules', ...String(name).split('/')), { recursive: true, force: true, maxRetries: 2 }); } catch {}
-    }
-    try { fs.rmSync(marker, { force: true }); } catch {}
-    log('boot', '已迁移到桌面专属 profile（' + DESKTOP_PROFILE + '）：旧 web profile 清理了 ' + cleaned.removed.length + ' 条桌面配套行 / ' + builtinNames.length + ' 个配套包');
-
-    // 3) 在专属 profile 里复活用户选择的皮肤（等 syncCompanionPlugins 写完
-    //    全部皮肤行之后执行，见 applyLegacySkinChoice）。
-    if (enabledSkin) {
-      const s2 = updater.loadSettings(updCtx());
-      s2.legacySkinChoice = enabledSkin;
-      updater.saveSettings(updCtx(), s2);
-      log('boot', '将迁移用户皮肤选择: ' + enabledSkin);
-    }
-  } catch (err) {
-    log('boot', '共享 profile 迁移失败（不影响启动）: ' + err.message);
-  }
-}
-
-function extractPatchRowIds(patch) {
-  const ids = [];
-  const re = /^\s*-\s*id:\s*([\w.-]+)\s*$/gm;
-  let m;
-  while ((m = re.exec(String(patch || ''))) !== null) ids.push(m[1]);
-  return ids;
-}
-
-// 按 id 集合删除 patch 里的 insert 行块（与 removeBundledRowDuplicates 同
-// 语法约定：id 紧跟 `- insert:` 之后）。
-function removePatchRowsById(patch, ids) {
-  const removed = [];
-  if (typeof patch !== 'string' || patch === '' || !ids || ids.size === 0) return { patch, removed };
-  const lines = patch.split(/\r?\n/);
-  const out = [];
-  for (let i = 0; i < lines.length; i++) {
-    const line = lines[i];
-    if (/^-\s*insert:/.test(line)) {
-      const mid = /^\s*-\s*id:\s*([\w.-]+)\s*$/.exec(lines[i + 1] || '');
-      if (mid && ids.has(mid[1])) {
-        removed.push(mid[1]);
-        let j = i + 1;
-        while (j < lines.length && !/^-\s*insert:/.test(lines[j]) && /^#/.test(lines[j]) === false && /^\s+\S/.test(lines[j])) j++;
-        i = j - 1;
-        continue;
-      }
-    }
-    out.push(line);
-  }
-  let text = out.join('\n').replace(/\n{3,}/g, '\n\n');
-  if (!text.endsWith('\n')) text += '\n';
-  return { patch: text, removed };
-}
-
-// syncCompanionPlugins 之后调用一次：把迁移带来的皮肤选择落到新 profile。
-function applyLegacySkinChoice() {
-  try {
-    const s = updater.loadSettings(updCtx());
-    const skin = s.legacySkinChoice;
-    if (!skin || !/^ui-skin-[\w-]+$/.test(skin)) return;
-    const patchFile = path.join(desktopProfileDir(), 'cordis.patch.yml');
-    if (!fs.existsSync(patchFile)) return;
-    const text = fs.readFileSync(patchFile, 'utf8');
-    const re = new RegExp('(- id: ' + skin + '\\b[^\\n]*\\n(?:      [^\\n]*\\n)*?)      disabled: true\\n');
-    const next = text.replace(re, '$1');
-    if (next !== text) {
-      fs.writeFileSync(patchFile, next);
-      log('boot', '已在专属 profile 启用迁移的皮肤: ' + skin);
-    }
-    delete s.legacySkinChoice;
-    updater.saveSettings(updCtx(), s);
-  } catch (err) {
-    log('boot', '应用迁移皮肤选择失败: ' + err.message);
-  }
-}
 
 
 // ---------------------------------------------------------------------------
