+++ dsh-desktop/plugin-updater.d.ts
@@ -1,0 +1,2 @@
+/** plugin-updater.d.ts — legacy `plugin-updater.js` 垫片。迁 TS 后删除。 */
+export declare function versionOfDir(dir: string): string | null;
