+++ dsh-desktop/test/patch-row-heal.test.mjs
@@ -108,11 +108,14 @@ test('dsh-soul-md schema: path 带默认值，不再是 required', () => {
 });
 
 // main.js 侧双保险：新增行必须显式写 config，且启动时 heal 存量坏行。
+// Task 5：COMPANION_PLUGINS 表与 syncCompanionPlugins 迁 lib/
+// plugin-registry-data.ts / lib/plugins.ts。
 test('main.js: soul-md 行带 config + 启动时执行存量 heal', () => {
-  const src = readFileSync(join(root, 'main.js'), 'utf8');
-  assert.match(src, /id:\s*'soul-md',[^\n]*config:\s*\{\s*path:\s*'soul\.md'\s*\}/);
-  assert.match(src, /healSoulMdPatchRow\(patch\)/);
-  assert.match(src, /block \+= configLinesFor\(p\.config\)/);
+  const registrySrc = readFileSync(join(root, 'lib', 'plugin-registry-data.ts'), 'utf8');
+  const pluginsSrc = readFileSync(join(root, 'lib', 'plugins.ts'), 'utf8');
+  assert.match(registrySrc, /id:\s*'soul-md',[^\n]*config:\s*\{\s*path:\s*'soul\.md'\s*\}/);
+  assert.match(pluginsSrc, /healSoulMdPatchRow\(patch\)/);
+  assert.match(pluginsSrc, /block \+= configLinesFor\(p\.config\)/);
 });
 
 // 市场安装（dsh plugin add 登记 bundles）与 overlay 写行双挂载 →
